fix(link): fix parse error with trailing whitespace in link tags
Due to ambiguity, a space is required between the "reference" and
"description" for the `@link` and `@linkplain` tags. Normally
tree-sitter doesn't require it (spaces are extras), but javadoc
references are highly ambiguous.

example: `{@link reference desc}`

Parsing would create an error for `{@link reference }`, because the
space itself was not optional: make it optional.

This fixes the most common parsing error across open source corpora. No
changes to node structure.